# docs/api/swagger.yaml
definitions:
...
  payload.AuthzPaths:
    additionalProperties:
      description: HTTP methods granted on a path. "*" means every method
      items:
        type: string
      type: array
    description: API path to the methods granted on it. A key of "*" grants every
      path
    type: object
  payload.AuthzPermissions:
    additionalProperties:
      $ref: '#/definitions/payload.AuthzSubjects'
    description: Permission set by category; today the only category is "users"
    type: object
  payload.AuthzSubjects:
    additionalProperties:
      $ref: '#/definitions/payload.AuthzPaths'
    description: Subject id to the paths granted to it
    type: object
...
  payload.GetAuthenticatedUserResponse:
    description: Response containing authenticated user details and their complete
      permission set
    properties:
      account:
        allOf:
        - $ref: '#/definitions/payload.UserResponse'
        description: Authenticated user's account information
      permissions:
        allOf:
        - $ref: '#/definitions/payload.AuthzPermissions'
        description: Complete set of permissions granted to the user through roles
          and policies
    type: object
  payload.HTTPMessage:
    description: HTTPMessage represents a message to be sent to the client trough
      HTTP REST API.
    properties:
      code:
        description: |-
          Code is a stable, machine-readable reason, present only where a client
          has to BRANCH on it. Message is prose and may be reworded; this is not.

          It exists because a 401 has two meanings a client must treat differently:
          an expired access token should be refreshed and the request retried,
          while a revoked one must not be — the refresh token was revoked in the
          same breath, so the retry burns two requests and fails anyway. Nothing in
          the old response distinguished them, and a client cannot be asked to
          match on English.

          Empty on every response that does not need one, so no existing client
          sees a change.
        example: token_revoked
        format: string
        type: string
      message:
        example: success
        format: string
        type: string
      method:
        example: GET
        format: string
        type: string
      path:
        example: /api/v1/users
        format: string
        type: string
      status_code:
        example: 200
        format: int32
        type: integer
      timestamp:
        example: "2021-07-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
  payload.UpdateMeRequest:
    description: Request payload for updating the authenticated user's profile information
      (all fields optional)
    properties:
      first_name:
        description: Updated first name
        example: John
        format: string
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        description: Updated last name
        example: Doe
        format: string
        maxLength: 50
        minLength: 2
        type: string
      password:
        description: New password (must meet security requirements)
        example: NewSecureP@ssw0rd
        format: password
        maxLength: 72
        minLength: 8
        type: string
    type: object
...
  payload.UserResponse:
    description: UserResponse represents a user entity.
    properties:
      admin:
        description: Indicates if the user has administrative privileges
        example: false
        type: boolean
      created_at:
        description: Timestamp when the user account was created
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      disabled:
        description: Indicates if the user account is currently disabled
        example: false
        type: boolean
      email:
        description: User's email address
        example: john.doe@example.com
        format: email
        type: string
      first_name:
        description: User's first name
        example: John
        format: string
        type: string
      id:
        description: Unique identifier for the user
        example: 019b4b0d-a682-7e82-a25a-b0671dc354c2
        format: uuid
        type: string
      last_name:
        description: User's last name
        example: Doe
        format: string
        type: string
      local_account:
        description: Indicates if this is a locally managed account (vs SSO/federated)
        example: true
        type: boolean
      updated_at:
        description: Timestamp when the user account was last updated
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
info:
  contact:
    email: info@goapitemplate.local
    name: API Support
    url: https://goapitemplate.local
  description: |-
    A production-shaped starting point for a Go HTTP REST API: hexagonal architecture, multi-tenant projects, RBAC through Open Policy Agent, database-backed rate limiting and resource limits, and OpenTelemetry traces and metrics throughout.

    `products` is the worked example entity -- a project-scoped CRUD resource that exercises every convention here. Copy it when adding your own; see docs/architecture/adding-an-entity.md.

    ### Authentication

    Four bearer credentials, each accepted by a different set of endpoints; see the security definitions below for which is which. Ordinary calls use **AccessToken**.

    A logout revokes the access token it was called with, not only the refresh token, so it stops working immediately. That check can be switched off (`authn.access.token.revocation.enabled`), and where it is, a logged-out access token keeps working until it expires -- which is why the lifetime is short.

    Refresh tokens are rotated: each refresh spends the token it was given and returns a successor, so a client **must** store what `POST /auth/refresh` returns. Presenting a spent one again reads as a replay and ends the session for whoever holds the live token too.

    ### Conventions

    - **Identifiers are UUID v7.** Every `id` in this API, without exception.
...

    ### Rate limiting

...

...

...

    ### Operations

...
  version: v1
paths:
...
  /me:
    get:
      description: Retrieve the profile information for the currently authenticated
        user
      operationId: 0199489b-f2f0-718a-a0cb-de8752ea864f
      produces:
      - application/json
      responses:
        "200":
          description: User information retrieved successfully
          schema:
            $ref: '#/definitions/payload.UserResponse'
        "400":
          description: Invalid request format or parameters
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "429":
          description: Too many requests -- RATE_LIMIT_EXCEEDED is the budget, RATE_LIMIT_UNAVAILABLE
            the limiter's own store
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      security:
      - AccessToken: []
      summary: Get authenticated user
      tags:
      - Me
    put:
      consumes:
      - application/json
      description: Update the profile information for the currently authenticated
        user
      operationId: 0199489b-f2f0-718e-a94d-b05a296eb818
      parameters:
      - description: User update request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.UpdateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "400":
          description: Invalid request format or validation failed
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "409":
          description: User already exists (duplicate email)
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "429":
          description: Too many requests -- RATE_LIMIT_EXCEEDED is the budget, RATE_LIMIT_UNAVAILABLE
            the limiter's own store
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      security:
      - AccessToken: []
      summary: Update authenticated user
      tags:
      - Me
  /me/authz:
    get:
      description: Retrieve the authorization details and permissions for the currently
        authenticated user
      operationId: 0199489b-f2f0-719e-b860-3b7ea6a86a1a
      produces:
      - application/json
      responses:
        "200":
          description: Authorization information retrieved successfully
          schema:
            $ref: '#/definitions/payload.GetAuthenticatedUserResponse'
        "400":
          description: Invalid request format or parameters
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "429":
          description: Too many requests -- RATE_LIMIT_EXCEEDED is the budget, RATE_LIMIT_UNAVAILABLE
            the limiter's own store
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      security:
      - AccessToken: []
      summary: Get authorization info
      tags:
      - Me
      - Authorization
...
securityDefinitions:
  AccessToken:
    description: 'The ordinary credential: `Bearer <access_token>`, from `POST /auth/login`
      or `POST /auth/refresh`. Every endpoint outside `/auth`, `/health` and `/version`
      takes it. A personal access token is presented the same way and is accepted
      anywhere an access token is -- which is how a script authenticates without a
      login. Deliberately short-lived; refresh rather than lengthening it.'
    in: header
    name: Authorization
    type: apiKey
  RefreshToken:
    description: '`Bearer <refresh_token>`, and accepted **only** by `POST /auth/refresh`
      and `DELETE /auth/logout`. Refreshing spends it and returns a successor: store
      what comes back. Presenting a spent one is treated as a replay and ends the
      whole session, including for whoever holds the live token.'
    in: header
    name: Authorization
    type: apiKey
  ResetPasswordToken:
    description: '`Bearer <reset_password_token>`, accepted only by `POST /auth/password/reset`.
      Issued by `POST /auth/password/recover` and delivered by email, single use.'
    in: header
    name: Authorization
    type: apiKey
  VerificationToken:
    description: '`Bearer <verification_token>`, accepted only by `POST /auth/verify/confirm`.
      It proves an email address rather than exercising a permission, so it is the
      one credential no authorization check is applied to -- the account it belongs
      to is still disabled at that point. It never travels in a URL: the emailed link
      points at the frontend, which presents the token in this header.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...